Start verb tests, add function cjr_get_verb_type to verb.{c,h}
I assume that using an enum may be better for figuring out what type of verb something is.
Therefore, this function will likely help many who want to use the library, so that they do not need to cast the string to an enum themselves.
The function cjr_get_ending will still return a string, however.